rfct(telegram): State storage moved to a separate function
Instead of each processor thread getting state in parallel,
the stateQueue function will call takeState. takeState then
checks if the state has been lent and not released.

Essentially what this means is that if someone sends 2
messages in sequence the first message update handler will
take the state and the second message update will have to
wait for the first message to release the state.

This state is the conversation state that is used to handle
messages in a special way e.g. when a command needs the user
to press a button or take some other sequence of actions.