fix: Removed unnecessary copy
There was a bug where query.Add(offset=N) would not set the correct
offset. As a precaution there was a copy of the update to make sure the
there was no issue with the same pointer value being sent into the queue
channel. The original issue was resolved but this copy variable stayed.