fix(telegram): No naked returns in doApiRequest and error retry limit in getUpdates
There was a bug where if the API would return an error, doApiRequest
will retry a few times and then the loop ends and `return nil, err`, but
err was nil so the app would crash because it thought err was nil so
there should be a valid pointer on the left side of the return.

This had revealed another bug where if the token was "" getUpdates would
spam the telegram server and got back the same error. So a retry limit
was added to getUpdates too.