feat: /addApiKey conv. state + GH api client (no persistance)
Added /addApiKey that transitions the state of the conversation into one
for processing an API key and which sends a message with the user's GH
username when a valid API key is set.

Conversation state storage is not yet persistant.

API key should get stored in a database, which will be implemented in
the future.